doc: add documentation
