+++ src/main/scala/it/unibo/shipps/controller/GameController.scala
@@ -1,16 +1,28 @@
 package it.unibo.shipps.controller
 
 import it.unibo.shipps.controller.GamePhase.{Battle, Positioning}
-import it.unibo.shipps.logic.BattleLogic
 import it.unibo.shipps.model.*
 import it.unibo.shipps.view.SimpleGui
 import it.unibo.shipps.view.renderer.ColorScheme
 
 import scala.swing.Swing
 
+/** Represents the different phases of the game.
+ * Positioning: Players place their ships on the board.
+ * Battle: Players take turns attacking each other's boards.
+ * GameOver: The game has ended, either by sinking all enemy ships or losing all own ships.
+ */
 enum GamePhase:
   case Positioning, Battle, GameOver
 
+/** Represents the state of the game
+ * @param board the player's board
+ * @param enemyBoard the enemy's board
+ * @param selectedShip the ship currently selected by the player
+ * @param gamePhase the current phase of the game
+ * @param attackResult a map of positions to their attack results
+ * @param cellColors a map of positions to their display colors
+ */
 case class GameState(
     board: PlayerBoard,
     enemyBoard: PlayerBoard,
@@ -58,6 +70,12 @@ case class GameState(
           cellColors = cellColors + (position     -> ColorScheme.SUNK)
         )
 
+/** Represents the controller for the game, managing the game state and interactions.
+ * @param initialBoard the board of the player
+ * @param enemyBoard the board of the enemy
+ * @param positioning the ship positioning logic
+ * @param view the GUI view for displaying the game state
+ */
 class GameController(
     initialBoard: PlayerBoard,
     enemyBoard: PlayerBoard,
@@ -100,6 +118,9 @@ class GameController(
       if state.gamePhase == GamePhase.Positioning then state.selectedShip else None
     ))
 
+  /** Handles the click on a cell based on the current game phase.
+   * @param pos the position of the cell clicked
+   */
   def onCellClick(pos: Position): Unit = {
     val newState = state.gamePhase match
       case GamePhase.Positioning => handleCellAction(state, pos)(positioning.moveShip)
@@ -115,12 +136,18 @@ class GameController(
     updateView()
   }
 
+  /** Handles the double click on a cell to rotate the ship during the positioning phase.
+   * @param pos the position of the cell double-clicked
+   */
   def onCellDoubleClick(pos: Position): Unit =
     if state.gamePhase == GamePhase.Positioning then
       val newState = handleCellAction(state, pos) { (board, ship, _) => positioning.rotateShip(board, ship) }
       state = newState
       updateView()
 
+  /** Handles the keyboard click to randomize ship positioning during the positioning phase.
+   * @param ships the list of ships to randomize
+   */
   def onKeyBoardClick(ships: List[Ship]): Unit =
     if state.gamePhase == GamePhase.Positioning then
       positioning.randomPositioning(PlayerBoard(), ships) match
@@ -130,6 +157,7 @@ class GameController(
         case Left(error) =>
           println("Error randomizing ships")
 
+  /** Handles the action when the game starts. */
   def onStartGame(): Unit = {
     val (newState, message) = state.gamePhase match
       case GamePhase.Positioning =>
