+++ src/main/scala/it/unibo/shipps/view/components/GridManager.scala
@@ -9,9 +9,16 @@ import it.unibo.shipps.view.renderer.ButtonRenderer
 import scala.swing.event.ButtonClicked
 import scala.swing.{Button, Swing}
 
+/** Manages the grid of buttons representing the game board. Handles button clicks and updates the game state.
+ * @param controller the game controller that manages the game logic
+ * */
 class GridManager(controller: GameController) {
   private var clickState = ClickState(None, 0L)
 
+  /** Updates the grid buttons based on the current game state.
+   * @param state the current game state
+   * @return a sequence of buttons representing the game board
+   */
   def createButtons(state: GameState): IndexedSeq[Button] = {
     for {
       y <- 0 until PlayerBoard.size
@@ -29,23 +36,6 @@ class GridManager(controller: GameController) {
           ClickHandler.handleClick(clickType, controller)
       }
       btn
-    }
-  }
-}
-
-object ButtonFactory {
-  def createGridButton(pos: Position, state: GameState): Button = {
-    new Button(ButtonRenderer.getText(pos, state)) {
-      opaque = true
-      border = Swing.LineBorder(java.awt.Color.LIGHT_GRAY)
-      background = ButtonRenderer.getColor(pos, state)
-    }
-  }
-
-  def createStartGameButton(): Button = {
-    new Button("Start Game") {
-      background = java.awt.Color.GREEN
-      foreground = java.awt.Color.BLACK
     }
   }
 }
