+++ src/main/scala/it/unibo/shipps/view/renderer/ColorScheme.scala
@@ -2,6 +2,9 @@ package it.unibo.shipps.view.renderer
 
 import java.awt.Color
 
+/** Color scheme for the game board.
+  * Defines colors for different states of the game board cells.
+  */
 object ColorScheme {
   val UNOCCUPIED: Color    = java.awt.Color.CYAN
   val SELECTED_SHIP: Color = java.awt.Color.YELLOW
